Initiali Implementation of HullModLogic.
