+++ css/HullMods/hullmods.css
@@ -31,6 +31,7 @@
 		border: 1px solid rgba(0, 0, 0, 0.5);
 
 		padding: 1.6rem;
+		gap: 1.6rem;
 
 		.hullmod {
 			/* width: max-content; */
